No more eviction.
Instead of keeping files open and moving the cursor along between intervals of
Tail, Franz uses IO::read so the file is only open during Tail::read. This is
more of a simplification of logic than anything, but it does marginally reduce
CPU and memory usage, and it nicely sidesteps the open file issue.